refactor: 테마 설정 코드 splash -> application으로 변경
